The new architecture is at a working point
Got the new architecture to a point where it works about the same as it
used to.  There are some tests that need to be run, and modules need to
be converted to the new system.  Debugging messages need to be added
basically everywhere, and I figure we'll just have them dump to file or
stdout for now.